move dpd-client dependency out of omicron common
This breaks downstreams that import omicron_common as the way dpd-client
gets built is specific to the omicron build machinery.
